refactor(local): split workflow orchestration by command
Keep LocalContext as the stable driving facade while moving project lifecycle, artifact queries, source-change workflows, and shared resolution into focused private modules with explicit dependencies.
